pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-optimized-rag"
version = "0.1.0"
description = "Zero-latency asynchronous memory router for voice agents using dual-agent architecture"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "VoiceAgentRAG Contributors"}]
keywords = ["rag", "voice", "real-time", "dual-agent", "memory", "cache"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.4",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
ollama = ["ollama>=0.3"]
gemini = ["google-genai>=1.0"]
qdrant = ["qdrant-client>=1.9"]
mcp = ["mcp>=1.6"]
voice = [
    "sounddevice>=0.4",
    "webrtcvad>=2.0.10",
    "faster-whisper>=1.0",
    "edge-tts>=6.1",
]
all = [
    "voice-optimized-rag[openai,anthropic,ollama,gemini,qdrant,mcp,voice]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://github.com/SalesforceAIResearch/VoiceAgentRAG"
Repository = "https://github.com/SalesforceAIResearch/VoiceAgentRAG"

[tool.setuptools.packages.find]
include = ["voice_optimized_rag*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
